adjust organization code location
